Deprecate de facto private methods
The introduction of the `Editor` class enables the deprecation of a
variety of class- and instance-level methods across the `ActionText`
namespace. The main motivation of these deprecations is to seize the
opportunity to reset Action Text's public API.

There are two categories of deprecations in this changeset:
`module`/`class` deprecations and method deprecations.

The `module` and `class` deprecations target modules that are internal
and "private" in intent, but public in Ruby and not marked as `:nodoc:`.
They're either particular to Trix, or their responsibilities have been
given to `ActionText::Editor`.

The method deprecations also target methods that are internal and
"private" in intent, but `public` in Ruby and not marked as `:nodoc:`.
In general, determining whether or not to deprecate a method hinged on
whether its name or arguments mentioned `"trix"`, *or* if it lacked
method-level documentation or comments of any kind.

When possible, the previous code paths have been deprecated *and*
re-implemented in terms of an `ActionText::TrixEditor` instance.