How to review this proposal
This exploration in expanding support for Action Text editors is motivated by [rails/actiontext#41](https://github.com/rails/actiontext/issues/41).

This diff consists of a sequence of commits that could be merged incrementally over time, but are presented here as a unit to form a more cohesive narrative, and to demonstrate a potential end result.

At the moment, the theme of the commits is as follows:

1. Add Promise support to Client-side JavaScript
2. Deprecate Trix-specific Ruby methods
3. Consolidate methods to the `ActionText::Editor` class, including a barebones demonstration of integrating with a third-party editor ([ProseMirror](https://prosemirror.net), in this case)
4. Remove the barebones demonstration

Motivation / Background
---

Related to [rails/actiontext#41](https://github.com/rails/actiontext/issues/41)

The main aim of this changeset is to provide a single, extensible entrypoint for third-party editors (that are not Trix) to integrate with Action Text.

The `Editor` class (and the `TrixEditor` implementation to integrate with [Trix](https://trix-editor.org)) is introduced without a public interface. They are both introduced marked with `:nodoc:` so that all of their methods are considered `:nodoc:` as well. They're the only "public" part of this PR at the moment. This diff acknowledges their existence, but doesn't commit to any particular interface.

Their surface area is intentionally small, and the scope of what's public is aggressively limited for the sake of deprecating the old methods and merging the changes that introduce the new methods. The idea is for Action Text to declare "public interface bankruptcy" for the majority of its inner classes. From that interface bankruptcy, third-party integrates can make ad-hoc extensions as they explore what methods are good candidates to be added and maintained as part of the `Editor` API (for all Ruby, HTTP, and JavaScript interfaces). Once there's a need to pave the cowpaths, Action Text will have an opportunity to make additions to the interface without much burden of responsibility to maintain historical choices.

Without the existence of the `Editor` class to consolidate the rich text processing and rendering canonicalized `<action-text-attachment>` elements, third-party extension would be challenging and potentially more tedious.

Detail
---

This proposes that the entry point should be the `ActionText::Editor` class. In addition to introducing the new class itself along with its first `ActionText::TrixEditor` subclass, this changeset also required additional integration points with the existing models and classes.

Additions
---

The initial responsibilities of the `ActionText::Editor` have been determined by consolidating a variety of (class- and instance-) methods across a variety of classes to reduce the number of touchpoints for applications and engines that provide third-party editor integrations. The vast majority of its method definitions are direct copy-and-pastes from their original sources. Any method that previously mentioned `trix` in its name or argument list has replaced and generalized that occurrence with `editor`.

For example, this proposal draws inspiration from the `ActiveStorage::Service` class and its related `Registry` and `Configurator` infrastructural classes to store per-record information about the editor used to create its constituent data.

To store that metadata, this commit introduces a new migration that adds an `action_text_rich_texts.editor_name` column to encode the editor. Since the out-of-the-box configuration (and the only historical choice) uses `trix`, the migration backfills old data with `action_text_rich_texts.editor_name = 'trix'`.

The editors will be configured through a new `config/rich_text.yml` file that also draws inspiration from `config/storage.yml`. Applications (as well as third-party Editor engines) will provide their own `config/rich_text.yml` files to introduce and configure additional editors.

Changes
---

With the addition of the `ActionText::Editor` class and the `ActionText::RichText#editor` method, there is an opportunity to outsource the details of rendering editor form controls along with filling them out in System Tests to the Editor itself. This commit inverts the responsibility to render the field by introducing the `Editor#rich_text_area_tag`. The `rich_text_area_tag` method accepts an instance of `ActionView::Base` as the `view_context` argument to access the necessary render helpers. Similarly, the
`Editor#fill_in_rich_text_area` method accepts a Capybara session argument.

There is an opportunity to define `Editor#rich_text_area_tag` in terms of `#render_in(view_context, **locals, &block)` once the `#render_in` interface is extended to incorporate call-site keyword arguments and blocks.

On the client-side, this changeset contains some noteworthy JavaScript modifications. For example, it exports the `AttachmentUpload` class so that third-party editors can utilize it. To do so, it also modifies the constructor arguments to be more generalized. Instead of expecting a `Trix.Attachment` instance and a browser `File`, it accepts a "delegate" object to provide the `File` and to respond to upload lifecycle methods.

Additionally, the `AttachmentUpload.start` function returns a `Promise` to match asynchronous programming styles that don't utilize callbacks.

Deprecations
---

The introduction of the `Editor` class enables the deprecation of a variety of class- and instance-level methods across the `ActionText` namespace. The main motivation of these deprecations is to seize the opportunity to reset Action Text's public API.

There are two categories of deprecations in this changeset: `module`/`class` deprecations and method deprecations.

The `module` and `class` deprecations target modules that are internal and "private" in intent, but public in Ruby and not marked as `:nodoc:`. They're either particular to Trix, or their responsibilities have been given to `ActionText::Editor`.

The method deprecations also target methods that are internal and "private" in intent, but `public` in Ruby and not marked as `:nodoc:`. In general, determining whether or not to deprecate a method hinged on whether its name or arguments mentioned `"trix"`, *or* if it lacked method-level documentation or comments of any kind.

When possible, the previous code paths have been deprecated *and* re-implemented in terms of an `ActionText::TrixEditor` instance.

Additional information
---

In its current state, this PR includes a test suite that `ProsemirrorEditor` implementation to integrate with the [ProseMirror](https://prosemirror.net) package. It contains a Ruby-side implementation as well as an ad hoc JavaScript implementation. The main purpose of this code was to explore what it'd take to replace Trix, and to drive out the necessary extensible methods and points of integration.

It is not my intent to foist the burden of maintaining a third-party editor plugin in the test suite. I am also not proposing that ProseMirror be the first third-party editor to be given built-in support. I'm not sure how to proceed with the code, and have included it here to provide a general sketch of how an application or engine might integrate a WYSIWYG editor. I've left them included in the diff to provide code reviewers from Core with an opportunity to assess the client-side requirements and scan for opportunity to introduce new abstractions that Rails would want to support as part of its "public" JavaScript API.

Once the rest of the code is in a mergable state, I'm happy to remove the Prosemirror Ruby / JavaScript / test code.